WIP - refactor training and evaluation pipelines
